release: v5.0.7 — /metrics endpoint, auth lockout, native unit tests, cppcheck CI
- /metrics: Prometheus text exposition (Basic auth) for Grafana dashboards
  outside Home Assistant — system, radar, network, security and CSI metrics
- per-IP brute-force lockout on web auth: 5 failed attempts/60 s -> 429 with
  Retry-After, exponential backoff 30 s -> 15 min, cleared by successful login
- first automated tests: new [env:native] + Unity, pio test -e native covers
  the lockout state machine and the metrics builder; wired into CI
- cppcheck static analysis in CI (tools/run_cppcheck.sh, same invocation
  locally); existing findings triaged: printf format casts, deleted copy
  ctors on buffer-owning services, documented inline suppressions
- html lang attribute follows the selected UI language (was hardcoded cs)
- docs: Prometheus metrics guide, Development & Testing section, release
  checklist and maintaining guide updated

+++ src/services/ConfigSnapshot.cpp
@@ -135,7 +135,7 @@ bool ConfigSnapshot::saveSnapshot(Preferences* prefs, const char* fwVersion, con
     if (_snapshotCount < SNAPSHOT_SLOTS) _snapshotCount++;
     saveMeta();
 
-    DBG("ConfigSnapshot", "Saved slot %d (%u bytes, fw=%s, reason=%s)", slot, written, fwVersion, reason);
+    DBG("ConfigSnapshot", "Saved slot %d (%u bytes, fw=%s, reason=%s)", slot, (unsigned)written, fwVersion, reason);
     return true;
 }
